Mock 動作を ResourcesController から MocksController に移動した
