+++ tests/TestCase/Controller/ResourcesControllerTest.php
@@ -1985,533 +1985,4 @@ class ResourcesControllerTest extends IntegrationTestCase
 
         $this->assertCount(9, $resources);
     }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_057_リソースが登録されていない場合、アクセスすると404のレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users');
-
-        $this->assertContentType('application/json');
-        $this->assertResponseCode(404);
-        $this->assertResponseEquals('{"status":404,"message":"Not Found"}');
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_058_GETのリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'GET',
-            'response' => json_encode([
-                'code' => 200,
-                'body' => [
-                    'status' => 200,
-                    'message' => 'OK',
-                ],
-            ]),
-        ];
-
-        $this->enableCsrfToken();
-        $this->enableSecurityToken();
-        $this->enableRetainFlashMessages();
-        $this->post('/resources/add', $data);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users');
-
-        $this->assertContentType('application/json');
-        $this->assertResponseCode(200);
-        $this->assertResponseEquals('{"status":200,"message":"OK"}');
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_059_POSTのリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'POST',
-            'response' => json_encode([
-                'code' => 200,
-                'body' => [
-                    'status' => 200,
-                    'message' => 'OK',
-                ],
-            ]),
-        ];
-
-        $this->enableCsrfToken();
-        $this->enableSecurityToken();
-        $this->enableRetainFlashMessages();
-        $this->post('/resources/add', $data);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'POST';
-        $this->post('/v1/users');
-
-        $this->assertContentType('application/json');
-        $this->assertResponseCode(200);
-        $this->assertResponseEquals('{"status":200,"message":"OK"}');
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_060_PUTのリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'PUT',
-            'response' => json_encode([
-                'code' => 200,
-                'body' => [
-                    'status' => 200,
-                    'message' => 'OK',
-                ],
-            ]),
-        ];
-
-        $this->enableCsrfToken();
-        $this->enableSecurityToken();
-        $this->enableRetainFlashMessages();
-        $this->post('/resources/add', $data);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'PUT';
-        $this->put('/v1/users');
-
-        $this->assertContentType('application/json');
-        $this->assertResponseCode(200);
-        $this->assertResponseEquals('{"status":200,"message":"OK"}');
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_061_DELETEのリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'DELETE',
-            'response' => json_encode([
-                'code' => 200,
-                'body' => [
-                    'status' => 200,
-                    'message' => 'OK',
-                ],
-            ]),
-        ];
-
-        $this->enableCsrfToken();
-        $this->enableSecurityToken();
-        $this->enableRetainFlashMessages();
-        $this->post('/resources/add', $data);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'DELETE';
-        $this->delete('/v1/users');
-
-        $this->assertContentType('application/json');
-        $this->assertResponseCode(200);
-        $this->assertResponseEquals('{"status":200,"message":"OK"}');
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_062_クエリパラメータ付きのリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users?q=abc',
-            'method' => 'GET',
-            'response' => json_encode([
-                'code' => 200,
-                'body' => [
-                    'status' => 200,
-                    'message' => 'OK',
-                ],
-            ]),
-        ];
-
-        $this->enableCsrfToken();
-        $this->enableSecurityToken();
-        $this->enableRetainFlashMessages();
-        $this->post('/resources/add', $data);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users?q=abc';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users?q=abc');
-
-        $this->assertContentType('application/json');
-        $this->assertResponseCode(200);
-        $this->assertResponseEquals('{"status":200,"message":"OK"}');
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_063_複数階層のリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        for ($i = 1; $i <= 10; $i++) {
-            $list = [];
-            for ($j = 1; $j <= $i; $j++) {
-                $list[] = substr('abcdefghijklmnopqrstuvwxyz', 0, mt_rand(1, 10));
-            }
-
-            $url = sprintf('/%s', implode('/', $list));
-
-            $data = [
-                'url' => $url,
-                'method' => 'GET',
-                'response' => json_encode([
-                    'code' => 200,
-                    'body' => [
-                        'status' => 200,
-                        'message' => 'OK',
-                    ],
-                ]),
-            ];
-
-            $this->loadModel('Resources');
-
-            /** @var \App\Model\Entity\Resource $resource */
-            $resource = $this->Resources->newEntity();
-            $resource = $this->Resources->patchEntity($resource, $data);
-            $this->Resources->save($resource);
-
-            $_SERVER['REQUEST_URI'] = $url;
-            $_SERVER['REQUEST_METHOD'] = 'GET';
-            $this->get($url);
-
-            $this->assertContentType('application/json');
-            $this->assertResponseCode(200);
-            $this->assertResponseEquals('{"status":200,"message":"OK"}');
-        }
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_064_クエリパラメータ付きの複数階層のリソースが登録されている場合、そのリソースにアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        for ($i = 1; $i <= 10; $i++) {
-            $list = [];
-            for ($j = 1; $j <= $i; $j++) {
-                $list[] = substr('abcdefghijklmnopqrstuvwxyz', 0, mt_rand(1, 10));
-            }
-
-            $url = sprintf('/%s?q=%d', implode('/', $list), $i);
-
-            $data = [
-                'url' => $url,
-                'method' => 'GET',
-                'response' => json_encode([
-                    'code' => 200,
-                    'body' => [
-                        'status' => 200,
-                        'message' => 'OK',
-                    ],
-                ]),
-            ];
-
-            $this->loadModel('Resources');
-
-            /** @var \App\Model\Entity\Resource $resource */
-            $resource = $this->Resources->newEntity();
-            $resource = $this->Resources->patchEntity($resource, $data);
-            $this->Resources->save($resource);
-
-            $_SERVER['REQUEST_URI'] = $url;
-            $_SERVER['REQUEST_METHOD'] = 'GET';
-            $this->get($url);
-
-            $this->assertContentType('application/json');
-            $this->assertResponseCode(200);
-            $this->assertResponseEquals('{"status":200,"message":"OK"}');
-        }
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_065_レスポンスが複数回登録されている場合、そのリソースに初回アクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'GET',
-            'response' => json_encode([
-                [
-                    'code' => 200,
-                    'body' => [
-                        'status' => 200,
-                        'message' => 'message 200',
-                    ],
-                ],
-                [
-                    'code' => 302,
-                    'body' => [
-                        'status' => 302,
-                        'message' => 'message 302',
-                    ],
-                ],
-                [
-                    'code' => 404,
-                    'body' => [
-                        'status' => 404,
-                        'message' => 'message 404',
-                    ],
-                ],
-            ]),
-            'count' => 0,
-        ];
-
-        $this->loadModel('Resources');
-
-        /** @var \App\Model\Entity\Resource $resource */
-        $resource = $this->Resources->newEntity();
-        $resource = $this->Resources->patchEntity($resource, $data);
-        $this->Resources->save($resource);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users');
-
-        $code = 200;
-        $this->assertContentType('application/json');
-        $this->assertResponseCode($code);
-        $this->assertResponseEquals(sprintf('{"status":%d,"message":"message %d"}', $code, $code));
-
-        $resource = $this->Resources->find()->first();
-        $this->assertSame(1, $resource->count);
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_066_レスポンスが3回分登録されている場合、そのリソースに2回めのアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'GET',
-            'response' => json_encode([
-                [
-                    'code' => 200,
-                    'body' => [
-                        'status' => 200,
-                        'message' => 'message 200',
-                    ],
-                ],
-                [
-                    'code' => 302,
-                    'body' => [
-                        'status' => 302,
-                        'message' => 'message 302',
-                    ],
-                ],
-                [
-                    'code' => 404,
-                    'body' => [
-                        'status' => 404,
-                        'message' => 'message 404',
-                    ],
-                ],
-            ]),
-            'count' => 1,
-        ];
-
-        $this->loadModel('Resources');
-
-        /** @var \App\Model\Entity\Resource $resource */
-        $resource = $this->Resources->newEntity();
-        $resource = $this->Resources->patchEntity($resource, $data);
-        $this->Resources->save($resource);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users');
-
-        $code = 302;
-        $this->assertContentType('application/json');
-        $this->assertResponseCode($code);
-        $this->assertResponseEquals(sprintf('{"status":%d,"message":"message %d"}', $code, $code));
-
-        $resource = $this->Resources->find()->first();
-        $this->assertSame(2, $resource->count);
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_067_レスポンスが3回分登録されている場合、そのリソースに3回めのアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'GET',
-            'response' => json_encode([
-                [
-                    'code' => 200,
-                    'body' => [
-                        'status' => 200,
-                        'message' => 'message 200',
-                    ],
-                ],
-                [
-                    'code' => 302,
-                    'body' => [
-                        'status' => 302,
-                        'message' => 'message 302',
-                    ],
-                ],
-                [
-                    'code' => 404,
-                    'body' => [
-                        'status' => 404,
-                        'message' => 'message 404',
-                    ],
-                ],
-            ]),
-            'count' => 2,
-        ];
-
-        $this->loadModel('Resources');
-
-        /** @var \App\Model\Entity\Resource $resource */
-        $resource = $this->Resources->newEntity();
-        $resource = $this->Resources->patchEntity($resource, $data);
-        $this->Resources->save($resource);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users');
-
-        $code = 404;
-        $this->assertContentType('application/json');
-        $this->assertResponseCode($code);
-        $this->assertResponseEquals(sprintf('{"status":%d,"message":"message %d"}', $code, $code));
-
-        $resource = $this->Resources->find()->first();
-        $this->assertSame(3, $resource->count);
-    }
-
-    /**
-     * @return void
-     * @throws \Exception
-     */
-    public function test_068_レスポンスが3回分登録されている場合、そのリソースに4回めのアクセスすると登録されたレスポンスを返却することができる()
-    {
-        $this->loadFixtures(
-            'Resources'
-        );
-
-        $data = [
-            'url' => '/v1/users',
-            'method' => 'GET',
-            'response' => json_encode([
-                [
-                    'code' => 200,
-                    'body' => [
-                        'status' => 200,
-                        'message' => 'message 200',
-                    ],
-                ],
-                [
-                    'code' => 302,
-                    'body' => [
-                        'status' => 302,
-                        'message' => 'message 302',
-                    ],
-                ],
-                [
-                    'code' => 404,
-                    'body' => [
-                        'status' => 404,
-                        'message' => 'message 404',
-                    ],
-                ],
-            ]),
-            'count' => 3,
-        ];
-
-        $this->loadModel('Resources');
-
-        /** @var \App\Model\Entity\Resource $resource */
-        $resource = $this->Resources->newEntity();
-        $resource = $this->Resources->patchEntity($resource, $data);
-        $this->Resources->save($resource);
-
-        $_SERVER['REQUEST_URI'] = '/v1/users';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-        $this->get('/v1/users');
-
-        $code = 200;
-        $this->assertContentType('application/json');
-        $this->assertResponseCode($code);
-        $this->assertResponseEquals(sprintf('{"status":%d,"message":"message %d"}', $code, $code));
-
-        $resource = $this->Resources->find()->first();
-        $this->assertSame(4, $resource->count);
-    }
 }
